classy approach is working for drawdown
